Move sketch to appropriately named subfolder
The Arduino IDE requires that a sketch be located in a folder of the same name. Although the name of the repository does match the sketch name, when GitHub's popular Clone or download > Download ZIP feature is used to download the contents of a repository the branch/tag/commit name is appended to the folder name, causing a mismatch.

When opening a file that does not meet this sketch/folder name matching requirement the Arduino IDE presents a dialog:

The file "arduino-canbus-monitor.ino" needs to be inside a sketch folder named "arduino-canbus-monitor".
Create this folder, move the file, and continue?

After clicking "OK" the Arduino IDE currently moves only the file arduino-canbus-monitor.ino to the new folder, leaving behind the other source files. This causes compilation of the sketch to fail:

arduino-canbus-monitor-master\arduino-canbus-monitor\arduino-canbus-monitor.ino:15:21: fatal error: mcp_can.h: No such file or directory

 #include "mcp_can.h"

                     ^